reworked structure to work with json and fetch
