+++ api/journal.json
@@ -1,0 +1,68 @@
+{
+  "journalEntries": [
+    {
+      "id": 1,
+      "date": "07/24/2025",
+      "concept": "HTML & CSS",
+      "entry": "We talked about HTML components and how to make grid layouts with Flexbox in CSS.",
+      "quote": "In the final analysis, a booger to one is not a booger to all. -George Duke-",
+      "mood": "Partly Sunny"
+    },
+    {
+      "id": 2,
+      "date": "07/31/27",
+      "concept": "Spacetime manipulation",
+      "entry": "We talked about manipulation of spacetime modalities via electrochemical crystalization pressure transference of inner quanta photonic emission.",
+      "quote": "It's all devistatingly true--except the bits that are lies --Douglas Adams--",
+      "mood": "Sunny"
+    },
+    {
+      "id": 3,
+      "date": "08/14/03",
+      "concept": "Flexbox",
+      "entry": "Found that the best kind of box is flexbox, cause it flexes. Then some one grid boxed me into oblivion.",
+      "quote": "C'mon man. Cornpop was a bad dude. --Joe Biden--",
+      "mood": "Partly Cloudy"
+    },
+    {
+      "id": 4,
+      "date": "11/4/16",
+      "concept": "Covfefe",
+      "entry": "Shrugs shoulders",
+      "quote": "Despite the negative press covfefe",
+      "mood": "Gloom and Doom"
+    },
+    {
+      "date": "07/24/2020",
+      "concept": "Coronka",
+      "entry": "We talked about outrageous made for tv pandemic scripts.",
+      "quote": "Eternal Salvation — Or TRIPLE your money back! -J.R. 'Bob' Dobbs-",
+      "mood": "Overcast",
+      "id": 5
+    },
+    {
+      "date": "01/01/0001",
+      "concept": "Time-Blank",
+      "entry": "Hail Eris! All Hail Discordia!",
+      "quote": "'Should' ain't 'is'. -J.R. 'Bob' Dobbs-",
+      "mood": "Overcast",
+      "id": 6
+    },
+    {
+      "date": "01/02/0001",
+      "concept": "Time-Bank",
+      "entry": "Hail Bob! All Hail Dobbscordia!",
+      "quote": "'Is' ain't 'should'. -J.R. 'Bob' Dobbs-",
+      "mood": "Partly Cloudy",
+      "id": 7
+    },
+    {
+      "date": "01/03/0001",
+      "concept": "Time-Bark",
+      "entry": "Hail Sirius! All Hail Dogscordia!",
+      "quote": "&ltsideways&gt&lthuge&gt&ltlaser&gtPshew!&lt/laser&gt&lt/huge&gt&lt/sideways&gt -A.I. 'Bob' Dobbs-",
+      "mood": "Partly Cloudy",
+      "id": 8
+    }
+  ]
+}
