--- scripts/main.js
@@ -1,2 +1,0 @@
-import { EntryListComponent } from "./JournalEntryList.js";
-EntryListComponent();
