+++ src/scripts/feed/JournalEntry.js
